Set pre-selected data in FormContext
