+++ static/js/src/advantage/distributor/components/DistributorShopForm/AddSubscriptions/AddSubscriptions.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect } from "react";
+import React, { useContext } from "react";
 import { Button, Col, Row, Select } from "@canonical/react-components";
 import { FormContext } from "../../../utils/FormContext";
 import {
@@ -7,7 +7,6 @@ import {
   generateUniqueId,
   Support as SupportEnum,
   SLA as SLAEnum,
-  getPreSelectedItem,
 } from "advantage/distributor/utils/utils";
 import SubscriptionCard from "./SubscriptionCard.tsx/SubscriptionCard";
 
@@ -17,10 +16,7 @@ const AddSubscriptions = () => {
     setProductType,
     subscriptionList,
     setSubscriptionList,
-    offer,
   } = useContext(FormContext);
-
-  const items = offer?.items ?? [];
 
   const handleProductTypeChange = (
     event: React.ChangeEvent<HTMLSelectElement>
@@ -53,19 +49,6 @@ const AddSubscriptions = () => {
     { label: "Ubuntu Pro Desktop", value: ProductTypes.desktop },
     { label: "Ubuntu Pro Virtual", value: ProductTypes.virtual },
   ];
-
-  useEffect(() => {
-    if (subscriptionList?.length === 0) {
-      const preSetItem = getPreSelectedItem(items);
-      items.length > 0 &&
-        preSetItem?.length > 0 &&
-        setSubscriptionList(preSetItem as SubscriptionItem[]);
-      localStorage.setItem(
-        "distributor-selector-subscriptionList",
-        JSON.stringify(preSetItem as SubscriptionItem[])
-      );
-    }
-  }, [offer]);
 
   return (
     <div data-testid="wrapper">
